+++ tests/diff/array_basics.rb
@@ -1,0 +1,14 @@
+a = [1, 2, 3]
+puts a.length
+puts a[0]
+puts a[2]
+puts a[-1]
+a << 4
+puts a.length
+puts a[3]
+a.push(5)
+puts a.length
+puts a.first
+puts a.last
+puts a.empty?
+puts [].empty?