+++ tests/diff/string_basics.rb
@@ -1,0 +1,11 @@
+a = "hello"
+b = "world"
+puts a + " " + b
+puts a.length
+puts (a + b).length
+puts a == "hello"
+puts a == "Hello"
+puts a == b
+puts ""
+puts "x".length
+puts "abc".to_s